--- app/test/page.js
@@ -1,6 +1,0 @@
-// eslint-disable-next-line import/no-unresolved
-import { TestForm } from "./testForm"
-
-export default function Test() {
-    return <TestForm></TestForm>
-}
